moves country selection/shuffling logic into quiz service
